move all params handling to deepquest/__init__.py
