+++ deepquest/__init__.py
@@ -1,4 +1,3 @@
-
 def train(config, changes={}):
     """
     Handles QE model training.
@@ -44,6 +43,9 @@ def score(files):
     """
     import deepquest.score
     score.main(files)
+
+
+#FIXME All the below parameters handling functions should be moved to somewhere more intuitive
 
 def set_gpu_id(gpuid):
     """
@@ -117,6 +119,61 @@ def setparameters(user_config_path):
         logger.exception("Exception occured: {}".format(exception))
 
     return parameters
+
+
+def update_parameters(params, updates, restrict=False):
+    """
+    Updates the parameters from params with the ones specified in updates
+    :param params: Parameters dictionary to update
+    :param updates: Updater dictionary
+    :param restrict: If True, parameters from the original dict are not overwritten.
+    :return:
+    """
+    from six import iteritems
+    for new_param_key, new_param_value in iteritems(updates):
+        if restrict and params.get(new_param_key) is not None:
+            params[new_param_key] = new_param_value
+        else:
+            params[new_param_key] = new_param_value
+
+    return params
+
+
+def compare_params(params_new, params_old, ignore=None):
+    """
+    Checks a new params dictionary against one from a previous model for differences.
+    :param params_new: new params dictionary
+    :param params_old: previous params dictionary
+    :param ignore: list of keys to ignore in comparison
+    """
+    import logging
+    logging.basicConfig(level=logging.INFO,
+                        format='[%(asctime)s] %(message)s', datefmt='%d/%m/%Y %H:%M:%S')
+    logger = logging.getLogger(__name__)
+    stop_flag = False
+    for key in params_old:
+        if key not in ignore:
+            if key not in params_new:
+                logger.info(
+                    'New config does not contain ' + key)
+                stop_flag = True
+            elif params_new[key] != params_old[key]:
+                logger.info('New model has ' + key + ': ' +
+                            str(params_new[key]) + ' but previous model has ' + key + ': ' + str(params_old[key]))
+                stop_flag = True
+    for key in params_new:
+        if (key not in params_old) and (key not in ignore):
+            logger.info('Previous config does not contain ' + key)
+            stop_flag = True
+    if stop_flag == True:
+        raise Exception('Model parameters not equal, can not resume training. ')
+    else:
+        logger.info(
+            'Previously trained config and new config are compatible. ')
+        return
+
+
+#FIXME Change default params to a global variable.
 
 def default_params(model='BiRNN'):
     """
